[CFG] Split determine_ad_stack_size into its own .cpp
determine_ad_stack_size and its ~20 file-local helpers are a
self-contained ~500-line algorithm (Tarjan SCC condensation +
fingerprint-deduplicated DP) that has no per-node CFGNode counterpart
and is not used by any other pass in control_flow_graph.cpp. Pulling it
into a sibling file drops control_flow_graph.cpp from ~2100 to ~1480
lines and lets a first-time reader see at a glance that the AD-stack
sizer is a distinct concern.

The build picks the file up automatically through the existing
`file(GLOB ... "quadrants/ir/*")` in cmake/QuadrantsCore.cmake; no
CMakeLists change required. The implementation is a verbatim move
(including the rewritten "What problem this solves" docstring); access
to `start_node` still works because the method is still defined on
ControlFlowGraph.
